requirements.txt
# Discord Bot Framework
discord.py==2.3.2

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1

# RSS Feed Parsing
feedparser==6.0.10

# AI APIs
openai==1.6.1
anthropic==0.8.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Environment Variables
python-dotenv==1.0.0

# Date/Time Handling
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# GitHub API
PyGithub==2.1.1

# Google Perspective API
google-api-python-client==2.111.0

# Async Support
asyncio==3.4.3
asyncpg==0.29.0

# Development Tools
pytest==7.4.4
pytest-asyncio==0.23.2
black==23.12.1
flake8==7.0.0